+++ src/test/java/org/chronotics/talaria/websocket/TestMessageQueueToOneSession.java
@@ -45,10 +45,10 @@ public class TestMessageQueueToOneSession {
     private static String topicUrl2 = "ws://localhost:8080/topic/?id="+id2;
     private static String topicUrl3 = "ws://localhost:8080/topic/?id="+id3;
     private static int port = 8080;
-    private static int awaitTimeOfClient = 1000; // ms
-    private static int startUpTimeOfClient = 1500; // ms
-    private static int startUpTimeOfServer = 1500; //ms
-    private static int stopTimeoutOfServer = 1000; // ms
+    private static long awaitTimeOfClient = 1000; // ms
+    private static long startUpTimeOfClient = 1500; // ms
+    private static long startUpTimeOfServer = 1500; //ms
+    private static long stopTimeoutOfServer = 1000; // ms
 
     private static JettyServer server = null;
